pyproject.toml
[tool.poetry]
name = "syphonmixer"
version = "0.1.0"
description = ""
authors = ["Will Harris <will@harris.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.3"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
